requirements.txt
astroid==2.2.5
autopep8==1.4.3
certifi==2019.3.9
Django~>2.1.5
django-cors-headers==2.4.0
django-filter==2.1.0
django-guardian==1.5.0
djangorestframework==3.9.0
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
pytz==2018.9
six==1.12.0
wrapt==1.11.1
django-extensions==2.1.6
pillow==6.1.0
psycopg2